test(mapgen): clarify Ecology test ownership
